Update patch set 1
Patch Set 1:

No I don't have a document.  I'll produce one if necessary, but haven't had time yet. 
Currently in cdt core there is an ICommandLauncherFactory extension point.  I guess Jeff created this extension point some time ago.  The ICommandLauncherFactory extension has these methods:

	public ICommandLauncher getCommandLauncher(IProject project);
	public ICommandLauncher getCommandLauncher(ICConfigurationDescription cfgd);
	public void registerLanguageSettingEntries(IProject project, List<? extends ICLanguageSettingEntry> entries);
	public List<ICLanguageSettingEntry> verifyLanguageSettingEntries(IProject project, List<ICLanguageSettingEntry> entries);

The org.eclipse.cdt.core.CommandLauncherManager actually loads/reads the extensions and uses them...i.e. gets a factory and calls its methods with a given project.

There is an existing implementation of this extension point in the docker launcher plugin (ContainerCommandLauncherFactory).  The impl of each of the 4 methods above looks for the managed builder config adapters, checks Jeff's managed builder enabled flag and if enabled returns a non-null ICommandLauncher (specifically, ContainerCommandLauncher).  This is implemented in the docker launcher plugin.

The org.eclipse.cdt.core.CommandLauncherManager loads/reads the extensions and uses them.

Rather than mix the core builder ICommandLauncherFactory with the ContainerCommandLauncherFactory (i.e. have it handle both the core build and the managed builder), I've created a second impl of ICommandLauncherFactory called CBuildCommandLauncherFactory.   What this impl does is 1) Look for ICBuildConfiguration adapter;  2) If found, look for (new) ICBuildCommandFactoryConfiguration adapter interface.   3) If found, call the corresponding getCommandLauncher(); registerLanguageSettingEntries, verifyLanguageSettingEntries methods on ICBuildCommandFactoryConfiguration adapter.   The CommandLauncherFactory basically just delegates the creation of the ICommandLauncher to the ICBuildCommandLauncherConfiguration impl (see impl below).   Because the core builder associates a single build configuration instance with a single project (via the ICBuildConfigurationProvider) and the ICBuildCommandFactoryConfiguration is an adapter on that project's build config the ICBuildCommandLauncherConfiguration will have access to the ICBuildConfiguration properties and/or toolchains within the given project's build configuration.

The CBuildCommandLauncherFactory is intended to function appropriately for any/all core-build based projects.   As discussed with Jeff on the bug, another way would be to put this delegation logic in the existing ContainerCommandLauncherFactory, but there are two things against this IMHO:  1) It would mix the code for both the managed builder and core builder into a single class (ContainerCommandLauncherFactory) and 2) This impl would then be bound to the docker launcher impl of ICommandLauncher and so have to be reimplemented for every new provider of ICommandLauncher (i.e. other than docker).   By having a second extension (CBuildCommandLauncherFactory) in core rather than in docker launcher and delegating to ICBuildCommandFactoryConfiguration, it can serve for any/all future core-build based projects.

The reason I added verifyLanguageSettingEntries and registerLanguageSettingEntries to ICBuildCommandLauncherConfiguration is that these two methods (without project param) exist on ICommandLauncherFactory.   They are required for managed builder, but I wasn't sure whether they would be required for core builder, so I added them.  If they are not required for core builder, then I believe it would make sense to simply remove them from ICBuildCommandLauncherConfiguration and have that interface with a single method:   ICommandLauncher getCommandLauncher();   I simply wasn't sure how the language settings were used in the core builder, and so added them for completeness.

In the case of CMake...or any other core-build based project, all they have to do is to create a build configuration that implements the ICBuildCommandLauncherConfiguration adapter...e.g.

public class CMakeBuildConfiguration extends CBuildConfiguration implements ICBuildCommandFactoryConfiguration {

...

public ICommandLauncher getCommandLauncher() {
// check properties, toolchains, etc for whether our ICommandLauncher is enabled
// if so return an instance of appropriate type...e.g.
    if (isEnabled())
        return new ContainerCommandLauncher();   // the ContainerCommandLauncher class is from docker launcher
    else return null;
}

All of the interaction with the CommandLauncherManager and ICommandLauncherFactory is essentially taken care of by this CBuildCommandLauncherFactory.  With Jeff's priority mechanism, it would be possible to create a new ICommandLauncherFactory if needed without changing the existing code (although to be avoided I would say).

This has no deps on docker specifically, rather it's all ICommandLauncher, ICommandLauncherFactory, ICBuildCommandLauncherConfiguration so no changes needed to any of this for remote, other types of ICommandLaunchers.

Patch-set: 1