Update patch set 3
Patch Set 3:

Patch Set 3 is a rebase of Patch Set 2 to apply to the latest master. It still seems to suffer from the same problem of the test suite hanging.

I triggered some jstack jobs to see what's going on:

https://hudson.eclipse.org/cdt/job/jstack-all/60/consoleFull
https://hudson.eclipse.org/cdt/job/jstack-all/61/consoleFull
https://hudson.eclipse.org/cdt/job/jstack-all/62/consoleFull
https://hudson.eclipse.org/cdt/job/jstack-all/63/consoleFull

It looks like in each case, the indexer is job waiting to acquire the PDOM write lock (and the test suite is waiting for the indexer). I'm not seeing any other threads that would be holding the PDOM write lock, so it seems we are somehow leaking the lock.

Patch-set: 3
Label: Verified=0