Update patch set 5
Patch Set 5:

There was a bug in the patch that caused an exception to be thrown on the semantic highlighting reconciler thread while running semantic highlighting tests.

I'm not sure how that gave rise to the PDOM write lock leaking - that probably deserves to be investigated and fixed independently - but fixing the bug fixed the tests hanging.

Patch-set: 5
Label: Verified=0