Restore removed pom.xmls for features and fix pom.xml versions
The pom.xmls are needed for features because the features live in
a different groupId than the plug-ins. This is needed because Eclipse
considers the namespace between features and plug-ins separately, but
Maven does not. As there are plug-ins and features in CDT with the same
name we use the groupId to differentiate them

Change-Id: Ia34ea37bcf0bfcda25a8e76a0f5b4a368d1e3465