Bug 512297 - Cache template argument list strings of template instances
This avoids quadratic performance in ASTTypeUtil on some code patterns.

Both normalized and unnormalized argument list strings are cached.

Currently, the cache is only used in the following places:

  - ASTTypeUtil.appendCppName(). Notably, this is used by
    TypeInstantiationRequest.hashCode(), which is a frequent
    performance problem.

  - CPPTemplates.haveSameArguments()

However, there are other places that build template argument list strings
that could transition to using the cache in the future.

Change-Id: I260877f820665cbe0939c0c3065514559592e721