Bug 525856 - LaunchBarManager logs exception during IProject.create
When creating a project with IProject.create, if the bundle
org.eclipse.launchbar.core is active, an error is seen in the error log.
This is caused by an exception thrown from
CBuildConfigurationManager.supports, since it will query the build
configurations of the project under creation (which is not accessible
yet).

This change ensures CBuildConfigurationManager.supports only asks for
project info if the project is a C/C++ project. This avoids the logged
exception.

Change-Id: Ifea6bbb72139f945b7828921b724159effee7be0
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>