Update patch set 1
Patch Set 1:

Thank you Nathan for suggesting a fix to this.
I would rather go for a ConcurrentHashMap to make the the stringCache threadsafe. It should be faster than locking on object level (see here, for example: http://crunchify.com/hashmap-vs-concurrenthashmap-vs-synchronizedmap-how-a-hashmap-can-be-synchronized-in-java/) and it does not clutter the code.

I would suggest this fix:

private final Map<Long, Reference<IString>> stringCache = new ConcurrentHashMap<>();

Patch-set: 1
Label: Code-Review=0