Update patch set 1
Patch Set 1:

I could not figure out how to change the CQ to point to the gerrit URL. It is strange though, as the exmaple in the "Contribution Record" is a link to bugzilla.

Eventually I withdrew it and created a new one [1]. Which was annoying as well, since the "Clone CQ" just redirected me to portal.eclipse.org.

[1] https://dev.eclipse.org/ipzilla/show_bug.cgi?id=14742

Patch-set: 1
Reviewer: Gerrit User 661 <661@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0