Update patch set 2
Patch Set 2:

> Uploaded patch set 3.

Ok, so another attempt. When changing the configuration I now remove the default one if it is there. This seems to work fine, except that when creating a new project you will get a "build/default" directory to build in, but then when you explicitly select "run" or "debug", the "build/default" directory is no longer used, but a new build directory "build/{run, debug}" is created, forcing a full rebuild.

Not sure where to go from here.

Patch-set: 2
Label: Verified=0