Update patch set 1
Patch Set 1:

I just re-ran the tests (including the one which failed on the Hudson CI) on my local machine and they all pass. Can someone look into this or run it again on the server to make sure it's really an issue on my codes part?

Patch-set: 1
Label: Verified=0