Update patch set 1
Patch Set 1:

Creating a ProblemType is good.

Having a checker is also good, because the ProblemType wouldn't be exposed by ProblemBindingChecker. (It may be worth adding a ProblemTypeChecker in the future to expose all ProblemTypes instead of having individual checkers like this one, but this is fine for now.)

As for a quick fix - do you think this is a common enough error for it to be worthwhile adding a quick fix? Do you know of any compiler that provides a quick fix for this error? I would lean towards not bothering with a quick fix.

In any case, if we do want to add a quick fix, we should add tests for it too.

Patch-set: 1
Label: Verified=0