Update patch set 2
Patch Set 2:

(8 comments)

Thank you for working on C++17 support!

In the future, I would prefer that separate features (e.g. in this case, "if with initializer" and "constexpr if") go into separate patches. This way, if one of the patches needs to be backed out for some reason (e.g. because it causes a regression), the other patch can remain. (There is no need to split this patch now, this is just a note for the future.)

Patch-set: 2
Label: Verified=0