Update patch set 6
Patch Set 6:

(1 comment)

The plan to remove the return type deductions stuff and commit the rest first sounds good to me. :) This looked a bit hacky to begin with to be honest so I'm happy if a better solution can be found.

I'll update this patch but I'm quite busy 'till february so it'll probably take a while.

Patch-set: 6
Label: Verified=0