Bug 530620 - Hook build of native serial port driver into maven build
In the same vein as commit dff6b3bae3d, this commit adds profiles
to trigger the build of the serial port component for the current
architecture, only when the profile is activated.

The serial port makefile is also improved to support building for
ppc64 and ppc64le architectures and to honour CFLAGS and LDFLAGS
that might be set in the environment, such as during RPM build.

I also hooked in the makefile "clean" targets so that invoking
"mvn clean verify" will do the right thing.

This simplifies the build for downstreams building from source
significantly.

Change-Id: Ie8cea95d953b256a1f62b77c76370ed7b3d81cee
Signed-off-by: Mat Booth <mat.booth@redhat.com>