Bug 530692 - Avoid EvalFunctionCall storing the evaluation for the implicit 'this' twice
EvalFunctionCall.fImplicitThis is sometimes redundant in that the
owner evaluation is already stored by one of the arguments. In
such cases, storing the owner separately in fImplicitThis can lead
to exponential complexity in chained method calls.

We resolve the duplication by computing the implicit this from the
function name evaluation instead of storing it where possible.

This was already implemented for cases where the function name
evaluation is an EvalMemberAccess in commit 659ff8c4a7c9. This
commit extends the approach to cases where the function name
evaluation is an EvalID.

Change-Id: Ic71e81b4692c51ffb8e15b3da9fc2dff1a554f05