Update patch set 1
Patch Set 1:

I had thought of adding it to the provider, but then if a user picks the toolchain, it won't work.  I'll make the change and open a separate bug for the quotes issue with that part of the pach.

Patch-set: 1
Reviewer: Gerrit User 582 <582@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0