Update patch set 1
Patch Set 1: Verified+1

I was able to verify that the exception goes away with this fix. With that said, this patch does not have a test case since I was not able to obtain a reproducer. How would you like to handle this? Should we submit this into master and keep an eye for regressions?

Patch-set: 1
Reviewer: Gerrit User 598 <598@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=+1