Update patch set 1
Patch Set 1:

Well, ideally you would have used org.eclipse.remote for the containers and used it's services architecture to get the toolchain and include path APIs you need :). But since you're not doing that, I'm not sure what the best approach is and will leave it at that.

Patch-set: 1