Update patch set 1
Patch Set 1: Code-Review-1

Thank you for looking at this, Felix!

However, this is not the right fix. To explain why, I'll start with some background information on what isCodanProblem() is for.

We support quick fixes for two different kinds of markers: (1) markers created by codan checkers, and (2) markers created for compiler errors by parsing the compiler's output. Sometimes, the same quick fix is applicable to both kinds of markers, and this is the case for the "Create local variable" quick fix.

isCodanProblem() is intended to distinguish between these two kinds of markers: it's supposed to return true for type (1), and false for type (2).

The problem of checking whether a variable resolution error detected by codan is in a local context, is already solved: it's done by ProblemBindingChecker annotating the marker with the argument ":func", and isApplicable() checking for that argument.

So the real bug is that isCodanProblem() is returning false when it should be returning true. The reason for this is that isCodanProblem() is returning a value which is only initialized in AbtractCodanCMarkerResolution.run(). That's fine when isCodanProblem() is called from apply() / modifyAst(), but it's a problem when it's called from isApplicable(), because run() hasn't been called by then.

A simple solution to this problem would be to have isCodanProblem() take the IMarker as an argument, and compute the answer (which is just "getProblemId(marker) != null") each time it's called, instead of looking up a cached value. This computation is not expensive, so I don't expect any performance impact.

We should also add a test case to CreateLocalVariableQuickFixTest for this. I note that the test harness for quick fixes (QuickFixTestCase base class) does not currently call isApplicable() at all, it just invokes the quick fix unconditionally. We should probably modify it to call isApplicable(), and to record its result, so we can make assertions about it.

(Finally, it's worth noting that this bug also exists, independently, for markers of type (2), for which isApplicable() always returns true. That can be solved using an approach similar to your current patch, possibly with some modifications (for example, I'm not sure whether getASTNameFromMarker() works for compiler error markers). I would suggest dealing with that in a separate bug, though.)

Patch-set: 1
Label: Code-Review=-1