Update patch set 1
Patch Set 1:

Thanks for the feedback!

With the background information you supplied, I can see how my fix is indeed not the right thing to do. I foolishly, without even checking further, assumed that, for some enigmatic reason, isCodanProblem() was not supposed to be called in isApplicable(IMarker) since its implementation did not fit in.

I believe that adjusting the API of isCodanProblem() to accept an IMarker would be the right thing to do, however, I fear that we would break a number of Codan plugins, some of which are possibly not even maintained anymore. As far as I can tell, the adjustments that would be required by CDT's own QuickFixes are minimal. I will try to catch Thomas later today, to see what he thinks about the issue.

With respect to the processing of IMarkers originating from compiler output, I will open an issue for that.

Finally, regarding the test harness in general, would you prefer If I were to open a separate issue to adjust it to check for isApplicable(IMarker) before executing each test case, or should bundle that change in with this patch?

Patch-set: 1
Label: Verified=0