Update patch set 1
Patch Set 1:

> I believe that adjusting the API of isCodanProblem() to accept an
 > IMarker would be the right thing to do, however, I fear that we
 > would break a number of Codan plugins, some of which are possibly
 > not even maintained anymore. 

Good point; I didn't realize AbstractCodanCMarkerResolution was public API.

What we can do in this case is keep around the zero-argument isCodanProblem(), marked as deprecated (with the current implementation in place), but also add a one-argument overload isCodanProblem(IMarker), and switch our internal usages (including the one at issue here) to the new overload.

 > Finally, regarding the test harness in general, would you prefer If
 > I were to open a separate issue to adjust it to check for
 > isApplicable(IMarker) before executing each test case, or should
 > bundle that change in with this patch?

We can do that in this patch.

Patch-set: 1