Update patch set 1
Patch Set 1:

It does look reasonable with the default light scheme and dark scheme.
In the light scheme the Contrast is a little reduced as the "Find Scope" Color defaults to (185,176,180) while the previous constant was (212,212,212).
The Find Scope Color will have a proper contrast on every nicely setup color scheme.

The only issue i could come up with is that it is not quiet intuitive to link the setting "Find Scope Color" to "Highlight Macro Expansion Step Color". But, as it is my first commit i did not dare to add a new setting...

Patch-set: 1
Reviewer: Gerrit User 177273 <177273@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0