Update patch set 4
Patch Set 4:

Thanks for the review!

Adjusting the harness to only run the QuickFix if applicable indeed exposed a bug in 'CaseBreakQuickFixFallthroughAttributeTest'! The issue being that the QuickFix to insert the attribute was not enabled (since it currently is disabled by default).

The typo in assertIsApplicableForAllMarkers was somewhat embarrassing considering I didn't write it late at night but during the day :)

Patch-set: 4
Reviewer: Gerrit User 110516 <110516@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0