Bug 532592: [tests] fix bitrotten GDBRemoteTracepointsTest
When I fixed the assume on "remote" in commit 4a447ce1a6dbd59000293810eddd75bac33f6f10
I exposed this test as having not been run in a while and it has
bitrotten due to changes in the compiler behaviour and the bad
definition of fastTracepointsSupported. This commit brings the test up
to speed.

Change-Id: I5eee6af4de111d864465851de91889aecf875dfb