Bug 396354: Extract Function incorrectly extracts goto statements
This patch fixes an exception which occurs when trying to extract a goto
statement. It further correctly handles such cases and adds tests
accordingly.

It also prepares the refactoring utilities for further improvements,
which will be part of patches building on top of this one.

Note: This patch is based on work done by Daniel Marty.

Change-Id: Ifc6a0787399aac8cc6c53840713b97fa577c2d7b
Signed-off-by: Hansruedi Patzen <hansruedi.patzen@hsr.ch>