Bug 330204 - Use ManagedBuilderManager to build
The UI is using ManagedBuildManager to build a specified set of
configurations.
Insead of duplicating the logic (since clean does not support argument
transfer from HeadlessBuilder all the way to CommonBuilder), call the
public function for building a set of configurations.

WARNING: This fix is relying on a race condition in
ManagedBuildManager.buildConfigurations() since that method swaps the
"active configuration" while invoking the builder for the clean target.

Change-Id: I422a22e43a0acbef85420c04028475d61ad2ff85
Signed-off-by: Torbjörn Svensson <torbjorn.svensson@st.com>