Bug 533766 - Deadlock at IDE shutdown with active debug session
When shutting down eclipse and there is an active debug session, eclipse
might deadlock if it can't destroy the debug session and would normally
leave a zombie process for the eclipse instance.  This fix allows
eclipse to properly shutdown if the debug session is destroyed in less
than 1 minute, else the IDE will simply be terminated.

Change-Id: Icb9b019c7ff2ec9cdc9870a392a657fe0dfde81b
Signed-off-by: Torbjörn Svensson <torbjorn.svensson@st.com>