Update patch set 1
Patch Set 1:

Thanks, I'm not sure what the rules are for updating the version number or not. I downloaded the oxygen 3a release, which contains org.eclipse.cdt.dsf.gdb 5.4.0 from 201802261533, so it might be ok to keep the current version number which is at 5.5.0. Not really sure here.

Patch-set: 1
Label: Verified=0