Update patch set 2
Patch Set 2:

> I think I did something wrong because it was supposed to be a
 > second patch set for https://git.eclipse.org/r/#/c/121404/

You can fix it by setting the "Change-Id" line in your commit message to match the one in 121404. That is how gerrit keeps them together. 

As there is already a fairly substantial review on an earlier version of this code in 121404, it might be best if you resubmitted this code into 121404 and abandoned this copy. That will make it easier for reviewers to see how you have progressed the code since the last review.

Thanks for working on the C++14 support. The whole CDT community will benefit from these improvements.

Patch-set: 2