Update patch set 2
Patch Set 2:

(2 comments)

Thanks, this is behaving better when I run it (no more semantic errors).

Your LexerTest is failing because the test harness is not passing in the fSupportDigitSeparators option.

To fix this, you can:

  - set CPP_OPTIONS.fSupportDigitSeparators near the top of the test file
  - pass CPP_OPTIONS as the second param to init() in the test

With regards to the FastCPartitionScanner, I see what you mean with sections of numbers (and sometimes all code until the next line) being colored blue. As that's still an improvement over the syntax errors we currently get, I'm happy to deal with that in a follow-up issue.

Patch-set: 2
Label: Code-Review=0