Update patch set 5
Patch Set 5: Code-Review+2

I agree with that, we should keep trying to get a proper fix in place. At least this solution means we will only need to update 1 place not 2. Took a look and couldn't see a reason not to use ManagedBuildManager.buildConfigurations here.

Thanks for the fix!

Patch-set: 5
Label: Code-Review=+2