Update patch set 1
Patch Set 1: Code-Review-1

IIRC, the only time this code is relevant only when a Launch node is selected.  In that instance, it is not possible to retrieve disassembly. 

I think my intention was to not force an update (i.e. clear visible disassembled code), but to disable certain actions.

As far as I can tell the refresh action is the only one that relies on there being a session id.  Would overriding the update method from AbstractDisassemblyAction in DisassemblyPart.ActionRefreshView to only look if the backend is available and has a debug context accomplish the same thing?

Patch-set: 1