Update patch set 1
Patch Set 1:

Thanks!

Would you also be willing to add a test for this in GCCBuildCommandParserTest.java?

I'm thinking we can just modify an existing test, like testAbstractBuildCommandParser_Basic(), to throw in an extra "-DFOO" into the test line, and then check for the presence of an added macro entry with the value 1.

Patch-set: 1
Label: Code-Review=0