Update patch set 1
Patch Set 1:

Hi Thomas,

Sorry that there have been so many build failures due to the standalone debugger. I have been re-triggering all those builds now that it is resolved, but this patchset is not failing for that reason but because of a compile error, so I have not retriggered the builds for this patchset.


08:19:41 [ERROR] Failed to execute goal org.eclipse.tycho:tycho-compiler-plugin:1.1.0:compile (default-compile) on project org.eclipse.cdt.core: Compilation failure: Compilation failure: 
08:19:41 [ERROR] /jobs/genie.cdt/cdt-verify-test-cdt-ui-only/workspace/core/org.eclipse.cdt.core/parser/org/eclipse/cdt/internal/core/dom/parser/cpp/GNUCPPSourceParser.java:[2289] 
08:19:41 [ERROR] 	ICPPASTStaticAssertDeclaration assertion = getNodeFactory().newStaticAssertion(expression, message);
08:19:41 [ERROR] 	                                                                               ^^^^^^^^^^
08:19:41 [ERROR] expression cannot be resolved to a variable

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0