Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Thomas,
> 
> Sorry that there have been so many build failures due to the standalone debugger. I have been re-triggering all those builds now that it is resolved, but this patchset is not failing for that reason but because of a compile error, so I have not retriggered the builds for this patchset.
> 

Hi Jonah

Thanks for retriggering the builds. I haven't recognized the compile error in the first place. It probably slipped in when I tried to reduce the modifications applied to the parser. I'll fix that and check with Hansruedi whether I missed some of his tests.

Patch-set: 1
Label: Verified=0