Update patch set 1
Patch Set 1: Code-Review+1

John, if you think the API changes are minor or unlikely to cause issues with downstream projects I would say lets get it in now, otherwise this won't make it out until September (or maybe even December if Sept is a bugfix only release).

I have had a look, and I don't think this new API is a problem, so +1 from me.

If you approve it, it should be announced on cdt-dev and get into master branch by Monday.

Patch-set: 1
Label: Code-Review=+1