Bug 535836: Log instead of pop-up errors in disassembly view
All sorts of views, variables, memory, etc can do a DSF operation
that causes an error. But for some reason only the disassembly
view generally displays this error to the user in a pop-up.

This change converts the pop-up (which says things like "Target not
available" or "Failed to execute MI command:
-var-create - * symbol_that_user_typed_that_doesnt_exist
Error message from debugger back end:
-var-create: unable to create variable object") to just log the
error instead.

Errors are dealt with in the UI already (example, bad location in
text box in toolbar gets an exclamation next to it). However, I don't
know if there are some errors that really do need to be displayed
to the user with a pop-up.

Change-Id: I27b398c34e558db98d34cef773a2689fa9222149
(cherry picked from commit ac2363cb925ceaa477699497c2f252d72f47f45d)